cloud_functions/harvest/projects/pyproject.toml
[tool.poetry]
name = "projects"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
data_pipeline_tools = { git = "https://github.com/tpximpact/GCP-Data-Pipelines.git", branch = "main" }
requests = "^2.28.2"
google-cloud-core = "^2.3.2"
google-cloud-bigquery = "^3.5.0"
google-cloud-secret-manager = "^2.15.1"
pyarrow = "^11.0.0"
pandas-gbq = "^0.19.1"
python-dateutil = "^2.8.2"
pyforecast = { git = 'https://github.com/urquha/py_forecast', branch = "master" }
holidays = "^0.21.13"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
